app/requirements.txt
google-cloud-bigquery 
streamlit
db-dtypes
plotly
pandas
numpy 
